定制化表 控制部分完成

+++ src/main/java/com/fate1412/crmSystem/module/customTable/service/ITableColumnDictService.java
@@ -33,5 +33,5 @@ public interface ITableColumnDictService extends IService<TableColumnDict>, MyBa
     
     JsonResult<?> updateByDTO(TableColumnUpdateDTO tableColumnUpdateDTO);
     
-    boolean delById(Long id);
+    boolean delById(Integer id);
 }
